pyproject.toml
[project]
name = "productanalysis"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13"


[tool.poetry.dependencies]
python = ">=3.13,<4.0"
flask = "^3.0.0"
flask-cors = "^6.0.1"
authlib = "^1.3.0"
python-dotenv = "^1.1.1"
requests = "^2.32.5"
pyjwt = "^2.10.1"
pyodbc = "^5.2.0"
pymssql = "^2.3.7"
polars = "^1.32.3"
pandas = "^2.3.2"
matplotlib = "^3.10.6"
seaborn = "^0.13.2"
altair = "^5.5.0"
pyarrow = "^21.0.0"
fsspec = "^2025.9.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
